# api/openapi.yaml
openapi: 3.0.0
info:
  title: CVRPTW Solver API
  description: API for solving vehicle routing problem (VRP) with loaders
  version: 1.0.0
  contact:
    name: Team 10

servers:
  - url: http://139.100.207.201:5000
    description: MVPv0
  - url: http://139.100.207.201:5001
    description: MVPv1
  - url: http://139.100.207.201:5011
    description: MVPv1.2
  - url: http://139.100.207.201:5002
    description: MVPv2
  - url: http://139.100.207.201:5022
    description: MVPv2.2

security:
  - Auth: []

paths:
  /solve:
    post:
      summary: Start solving VRP
      description: |
        Accepts a CVRPTW scenario and starts solving in background.
        Returns immediately with status "started". Poll /solution for result.
      operationId: startSolving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scenario'
            examples:
              simple:
                summary: сценарий 1
                value:
                  vehicle_capacity: 1000
                  vehicle_speed: 3
                  loader_speed: 1
                  vehicle_shift_size: 720
                  loader_shift_size: 720
                  depot:
                    id: 0
                    x: 100
                    y: 100
                    load_time: 40
                  orders:
                    - id: 1
                      x: 134
                      y: 6
                      volume: 20
                      time_window: [988, 1031]
                      vehicle_service_time: 10
                      loader_cnt: 0
                      loader_service_time: 0
                      optional: 0
                  weights:
                    vehicle_salary: 380
                    loader_salary: 230
                    fuel_cost: 2
                    loader_work: 1
                    optional_order_penalty: 164
      responses:
        '202':
          description: Computation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: started
        '409':
          description: Already computing
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: computing
                  message:
                    type: string
                    example: Already solving
        '400':
          description: Invalid request / Scenario cannot be solved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /solution:
    get:
      summary: Get computed solution
      description: |
        Returns the solution if computation is done,
        or current status (computing / idle / error).
      operationId: getSolution
      responses:
        '200':
          description: Solution or status
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      status:
                        type: string
                        example: done
                      solution:
                        $ref: '#/components/schemas/Solution'
                  - type: object
                    properties:
                      status:
                        type: string
                        example: computing
                      stage:
                        type: string
                        example: solving_vehicles
                  - type: object
                    properties:
                      status:
                        type: string
                        example: idle
                      message:
                        type: string
        '500':
          description: Computation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string

  /metrics:
    get:
      summary: Get route metrics
      description: |
        Returns cost breakdown and summary metrics for the computed solution,
        or current status (computing / idle / error).
      operationId: getMetrics
      responses:
        '200':
          description: Metrics or status
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      status:
                        type: string
                        example: done
                      metrics:
                        $ref: '#/components/schemas/Metrics'
                  - type: object
                    properties:
                      status:
                        type: string
                        example: computing
                      stage:
                        type: string
                        example: solving_vehicles
                  - type: object
                    properties:
                      status:
                        type: string
                        example: idle
                      message:
                        type: string
        '500':
          description: Computation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string

  /validate:
    post:
      summary: Validate input scenario
      description: |
        Accepts a CVRPTW scenario and validates it without solving.
        Returns ok or a list of validation errors.
      operationId: validateScenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scenario'
      responses:
        '200':
          description: Input is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                          example: depot.x
                        message:
                          type: string

  /health:
    get:
      summary: Service health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /history:
    get:
      summary: List calculation history
      description: |
        Returns a list of past calculations with their key metadata
        (ID, timestamp, execution time, objective function cost, status).
      operationId: listHistory
      security:
        - Auth: []
      responses:
        '200':
          description: List of past calculations
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /history/{calculation_id}:
    get:
      summary: Get calculation details
      description: |
        Returns full details of a specific past calculation,
        including the original input data and the resulting output
        solution (or error details).
      operationId: getHistoryDetail
      security:
        - Auth: []
      parameters:
        - name: calculation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Calculation details with input/output data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryDetail'
        '404':
          description: Calculation not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Not found
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  schemas:
    Scenario:
      type: object
      required:
        - vehicle_capacity
        - vehicle_speed
        - loader_speed
        - vehicle_shift_size
        - loader_shift_size
        - depot
        - orders
        - weights
      properties:
        vehicle_capacity:
          type: number
          description: Vehicle capacity
          example: 1000
        vehicle_speed:
          type: number
          description: Vehicle speed (distance/time)
          example: 3
        loader_speed:
          type: number
          description: Loader speed
          example: 1
        vehicle_shift_size:
          type: number
          description: Vehicle shift duration
          example: 720
        loader_shift_size:
          type: number
          description: Loader shift duration
          example: 720
        depot:
          $ref: '#/components/schemas/Depot'
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        weights:
          $ref: '#/components/schemas/Weights'

    Depot:
      type: object
      required:
        - id
        - x
        - y
        - load_time
      properties:
        id:
          type: integer
          example: 0
        x:
          type: number
          example: 100
        y:
          type: number
          example: 100
        load_time:
          type: number
          description: Loading time
          example: 40

    Order:
      type: object
      required:
        - id
        - x
        - y
        - volume
        - time_window
        - vehicle_service_time
        - loader_cnt
        - loader_service_time
        - optional
      properties:
        id:
          type: integer
          example: 1
        x:
          type: number
          example: 134
        y:
          type: number
          example: 6
        volume:
          type: number
          example: 20
        time_window:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
          example: [988, 1031]
        vehicle_service_time:
          type: number
          description: Vehicle service time per order
          example: 10
        loader_cnt:
          type: integer
          description: Required number of loaders
          example: 0
        loader_service_time:
          type: number
          description: Loader service time per order
          example: 0
        optional:
          type: boolean
          description: Whether the order is optional
          example: 0

    Weights:
      type: object
      properties:
        vehicle_salary:
          type: number
          example: 380
        loader_salary:
          type: number
          example: 230
        fuel_cost:
          type: number
          example: 2
        loader_work:
          type: number
          example: 1
        optional_order_penalty:
          type: number
          example: 164

    Solution:
      type: object
      required:
        - loaders
        - vehicles
      properties:
        loaders:
          type: array
          items:
            $ref: '#/components/schemas/LoaderRoute'
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/VehicleRoute'

    VehicleRoute:
      type: object
      required:
        - id
        - route
        - time
      properties:
        id:
          type: integer
          description: Vehicle ID
          example: 1
        route:
          type: array
          items:
            type: integer
          description: Route (0 - depot, positive - order IDs)
          example: [0, 1, 2, 3, 0]
        time:
          type: array
          items:
            type: number
          description: Start time of operations at each route point
          example: [0, 100, 200, 300, 400]

    LoaderRoute:
      type: object
      required:
        - id
        - route
      properties:
        id:
          type: integer
          description: Loader ID
          example: 1
        route:
          type: array
          items:
            type: integer
          description: Loader route (order IDs)
          example: [1, 2, 3]

    Metrics:
      type: object
      properties:
        total_cost:
          type: number
          example: 1234.56
        fuel_cost:
          type: number
          example: 500.00
        vehicle_salaries:
          type: number
          example: 380.00
        loader_salaries:
          type: number
          example: 230.00
        loader_work_cost:
          type: number
          example: 120.00
        penalties:
          type: number
          example: 0
        vehicles_used:
          type: integer
          example: 3
        loaders_used:
          type: integer
          example: 2

    HistoryRecord:
      type: object
      properties:
        calculation_id:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: "2026-07-10T12:00:00+00:00"
        execution_time:
          type: number
          nullable: true
          description: Duration of calculation in seconds
          example: 12.34
        objective_function_cost:
          type: number
          nullable: true
          description: Total cost of the solution
          example: 1234.56
        status:
          type: string
          enum: [processing, success, error]
          example: success

    HistoryDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/HistoryRecord'
        - type: object
          properties:
            input_json_path:
              type: string
              description: Path to the saved input JSON file
            output_json_path:
              type: string
              nullable: true
              description: Path to the saved output JSON file
            input:
              $ref: '#/components/schemas/Scenario'
              nullable: true
            output:
              oneOf:
                - $ref: '#/components/schemas/Solution'
                - type: object
                  properties:
                    error:
                      type: string
                  description: Error details if status is error
              nullable: true

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request body
        message:
          type: string
          example: Missing required field 'vehicle_capacity'
        code:
          type: integer
          example: 400

  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid or missing API key

  securitySchemes:
    Auth:
      type: apiKey
      in: header
      name: X-API-Key